+++ packages/terra-core-docs/CHANGELOG.md
@@ -4,8 +4,11 @@
 
 * Changed
   * Fixed Focus Issue & Expand/collapse state for `terra-toggle` examples.
+
 * Added 
   * Added Label Example for `terra-toggle` examples.
+  * Added A11y guide for terra-icon.
+  * Added examples and accessibility guide for `terra-image`.
 
 ## 1.18.0 - (January 9, 2023)